fix: features chore: made tests seperate and removed clippy for temp
